[Feature #163136029] refactor rsvp response validation
